day-15-pr-review-agent/pyproject.toml
[project]
name = "day-15-pr-review-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.133.0",
    "httpx>=0.28.1",
    "langchain-openai>=0.2.14",
    "langgraph>=1.0.1",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.41.0",
]
